Revert remote storage, move all utils to storage.py

Remove fs_utils
