+++ python/ray/air/_internal/remote_storage.py
@@ -1,17 +1,24 @@
+import fnmatch
 import os
+import pathlib
 import sys
 import time
 import urllib.parse
 from pathlib import Path
 from pkg_resources import packaging
 import psutil
+import shutil
 from typing import Any, Dict, List, Optional, Tuple
+
+from ray.air._internal.filelock import TempFileLock
 
 try:
     import fsspec
+    from fsspec.implementations.local import LocalFileSystem
 
 except ImportError:
     fsspec = None
+    LocalFileSystem = object
 
 try:
     import pyarrow
@@ -20,9 +27,76 @@ try:
 except (ImportError, ModuleNotFoundError):
     pyarrow = None
 
+from ray import logger
+
 
 # Re-create fs objects after this amount of seconds
 _CACHE_VALIDITY_S = 300
+
+
+class _ExcludingLocalFilesystem(LocalFileSystem):
+    """LocalFileSystem wrapper to exclude files according to patterns.
+
+    Args:
+        exclude: List of patterns that are applied to files returned by
+            ``self.find()``. If a file path matches this pattern, it will
+            be excluded.
+
+    """
+
+    def __init__(self, exclude: List[str], **kwargs):
+        super().__init__(**kwargs)
+        self._exclude = exclude
+
+    @property
+    def fsid(self):
+        return "_excluding_local"
+
+    def _should_exclude(self, name: str) -> bool:
+        """Return True if `name` matches any of the `self._exclude` patterns."""
+        alt = None
+        if os.path.isdir(name):
+            # If this is a directory, also test it with trailing slash
+            alt = os.path.join(name, "")
+        for excl in self._exclude:
+            if fnmatch.fnmatch(name, excl):
+                return True
+            if alt and fnmatch.fnmatch(alt, excl):
+                return True
+        return False
+
+    def find(self, path, maxdepth=None, withdirs=False, detail=False, **kwargs):
+        """Call parent find() and exclude from result."""
+        names = super().find(
+            path, maxdepth=maxdepth, withdirs=withdirs, detail=detail, **kwargs
+        )
+        if detail:
+            return {
+                name: out
+                for name, out in names.items()
+                if not self._should_exclude(name)
+            }
+        else:
+            return [name for name in names if not self._should_exclude(name)]
+
+
+def _pyarrow_fs_copy_files(
+    source, destination, source_filesystem=None, destination_filesystem=None, **kwargs
+):
+    if isinstance(source_filesystem, pyarrow.fs.S3FileSystem) or isinstance(
+        destination_filesystem, pyarrow.fs.S3FileSystem
+    ):
+        # Workaround multi-threading issue with pyarrow
+        # https://github.com/apache/arrow/issues/32372
+        kwargs.setdefault("use_threads", False)
+
+    return pyarrow.fs.copy_files(
+        source,
+        destination,
+        source_filesystem=source_filesystem,
+        destination_filesystem=destination_filesystem,
+        **kwargs,
+    )
 
 
 def _assert_pyarrow_installed():
@@ -287,7 +361,7 @@ def get_fs_and_path(
     if is_local_path(uri) and not scheme:
         # Append local filesys scheme such that the downstream operations work
         # properly on Linux and Windows.
-        uri = "file://" + Path(uri).as_posix()
+        uri = "file://" + pathlib.Path(uri).as_posix()
 
     parsed = urllib.parse.urlparse(uri)
     # for uri="hdfs://48bb8ca83706:8020/test":
@@ -338,6 +412,23 @@ def get_fs_and_path(
     return None, None
 
 
+def delete_at_uri(uri: str):
+    _assert_pyarrow_installed()
+
+    fs, bucket_path = get_fs_and_path(uri)
+    if not fs:
+        raise ValueError(
+            f"Could not clear URI contents: "
+            f"URI `{uri}` is not a valid or supported cloud target. "
+            f"Hint: {fs_hint(uri)}"
+        )
+
+    try:
+        fs.delete_dir(bucket_path)
+    except Exception as e:
+        logger.warning(f"Caught exception when clearing URI `{uri}`: {e}")
+
+
 def read_file_from_uri(uri: str) -> bytes:
     _assert_pyarrow_installed()
 
@@ -353,68 +444,147 @@ def read_file_from_uri(uri: str) -> bytes:
         return file.read()
 
 
-def download_from_uri(
-    uri: str,
-    local_path: str,
-    filelock: bool = True,
-):
-    source_filesystem, source_path = get_fs_and_path(uri)
-    if not source_filesystem:
+def download_from_uri(uri: str, local_path: str, filelock: bool = True):
+    """Downloads a directory or file at a URI to a local path.
+
+    If the URI points to a directory, then the full directory contents are
+    copied, and `local_path` is the downloaded directory.
+    If the download fails, the `local_path` contents are
+    cleaned up before raising, if the directory did not previously exist.
+    NOTE: This creates `local_path`'s parent directories if they do not
+    already exist. If the download fails, this does NOT clean up all the parent
+    directories that were created.
+
+    Args:
+        uri: The URI to download from.
+        local_path: The local path to download to.
+        filelock: Whether to require a file lock before downloading, useful for
+            multiple downloads to the same directory that may be happening in parallel.
+
+    Raises:
+        ValueError: if the URI scheme is not supported.
+        FileNotFoundError: if the URI doesn't exist.
+    """
+    _assert_pyarrow_installed()
+
+    fs, bucket_path = get_fs_and_path(uri)
+    if not fs:
         raise ValueError(
-            f"Could not upload to URI: "
+            f"Could not download from URI: "
             f"URI `{uri}` is not a valid or supported cloud target. "
             f"Hint: {fs_hint(uri)}"
         )
 
-    from ray.train._internal.fs_utils import _download_from_fs_path
+    _local_path = Path(local_path).resolve()
+    exists_before = _local_path.exists()
+    if is_directory(uri):
+        _local_path.mkdir(parents=True, exist_ok=True)
+    else:
+        _local_path.parent.mkdir(parents=True, exist_ok=True)
 
-    _download_from_fs_path(
-        fs=source_filesystem,
-        fs_path=source_path,
-        local_path=local_path,
-        filelock=filelock,
-    )
+    try:
+        if filelock:
+            with TempFileLock(f"{os.path.normpath(local_path)}.lock"):
+                _pyarrow_fs_copy_files(bucket_path, local_path, source_filesystem=fs)
+        else:
+            _pyarrow_fs_copy_files(bucket_path, local_path, source_filesystem=fs)
+    except Exception as e:
+        # Clean up the directory if downloading was unsuccessful.
+        if not exists_before:
+            shutil.rmtree(local_path, ignore_errors=True)
+        raise e
 
 
 def upload_to_uri(
-    local_path: str,
-    uri: str,
-    exclude: Optional[List[str]] = None,
+    local_path: str, uri: str, exclude: Optional[List[str]] = None
 ) -> None:
-    destination_filesystem, destination_path = get_fs_and_path(uri)
-    if not destination_filesystem:
+    """Uploads a local directory or file to a URI.
+
+    NOTE: This will create all necessary directories at the URI destination.
+
+    Args:
+        local_path: The local path to upload.
+        uri: The URI to upload to.
+        exclude: A list of filename matches to exclude from upload. This includes
+            all files under subdirectories as well.
+            Ex: ["*.png"] to exclude all .png images.
+
+    Raises:
+        ValueError: if the URI scheme is not supported.
+    """
+    _assert_pyarrow_installed()
+
+    fs, bucket_path = get_fs_and_path(uri)
+    if not fs:
         raise ValueError(
             f"Could not upload to URI: "
             f"URI `{uri}` is not a valid or supported cloud target. "
             f"Hint: {fs_hint(uri)}"
         )
 
-    from ray.train._internal.fs_utils import _upload_to_fs_path
+    if not exclude:
+        _ensure_directory(bucket_path, fs=fs)
+        _pyarrow_fs_copy_files(local_path, bucket_path, destination_filesystem=fs)
+    elif fsspec:
+        # If fsspec is available, prefer it because it's more efficient than
+        # calling pyarrow.fs.copy_files multiple times
+        _upload_to_uri_with_exclude_fsspec(
+            local_path=local_path, fs=fs, bucket_path=bucket_path, exclude=exclude
+        )
+    else:
+        # Walk the filetree and upload
+        _upload_to_uri_with_exclude_pyarrow(
+            local_path=local_path, fs=fs, bucket_path=bucket_path, exclude=exclude
+        )
 
-    _upload_to_fs_path(
-        local_path=local_path,
-        fs=destination_filesystem,
-        fs_path=destination_path,
-        exclude=exclude,
+
+def _upload_to_uri_with_exclude_fsspec(
+    local_path: str, fs: "pyarrow.fs", bucket_path: str, exclude: Optional[List[str]]
+) -> None:
+    local_fs = _ExcludingLocalFilesystem(exclude=exclude)
+    handler = pyarrow.fs.FSSpecHandler(local_fs)
+    source_fs = pyarrow.fs.PyFileSystem(handler)
+
+    _ensure_directory(bucket_path, fs=fs)
+    _pyarrow_fs_copy_files(
+        local_path, bucket_path, source_filesystem=source_fs, destination_filesystem=fs
     )
 
 
-def delete_at_uri(uri: str):
-    fs, fs_path = get_fs_and_path(uri)
-    if not fs:
-        raise ValueError(
-            f"Could not clear URI contents: "
-            f"URI `{uri}` is not a valid or supported cloud target. "
-            f"Hint: {fs_hint(uri)}"
-        )
+def _upload_to_uri_with_exclude_pyarrow(
+    local_path: str, fs: "pyarrow.fs", bucket_path: str, exclude: Optional[List[str]]
+) -> None:
+    def _should_exclude(candidate: str) -> bool:
+        for excl in exclude:
+            if fnmatch.fnmatch(candidate, excl):
+                return True
+        return False
 
-    from ray.train._internal.fs_utils import _delete_fs_path
+    for root, dirs, files in os.walk(local_path):
+        rel_root = os.path.relpath(root, local_path)
+        for file in files:
+            candidate = os.path.join(rel_root, file)
 
-    _delete_fs_path(fs, fs_path)
+            if _should_exclude(candidate):
+                continue
+
+            full_source_path = os.path.normpath(os.path.join(local_path, candidate))
+            full_target_path = os.path.normpath(os.path.join(bucket_path, candidate))
+
+            _ensure_directory(str(Path(full_target_path).parent), fs=fs)
+            _pyarrow_fs_copy_files(
+                full_source_path, full_target_path, destination_filesystem=fs
+            )
 
 
 def list_at_uri(uri: str) -> List[str]:
-    fs, fs_path = get_fs_and_path(uri)
+    """Returns the list of filenames at a URI (similar to os.listdir).
+
+    If the URI doesn't exist, returns an empty list.
+    """
+    _assert_pyarrow_installed()
+
+    fs, bucket_path = get_fs_and_path(uri)
     if not fs:
         raise ValueError(
             f"Could not list at URI: "
@@ -424,29 +594,50 @@ def list_at_uri(uri: str) -> List[str]:
 
     if not is_non_local_path_uri(uri):
         # Make sure local paths get expanded fully
-        fs_path = os.path.abspath(os.path.expanduser(fs_path))
+        bucket_path = os.path.abspath(os.path.expanduser(bucket_path))
 
-    from ray.train._internal.fs_utils import _list_at_fs_path
-
-    return _list_at_fs_path(fs, fs_path)
+    selector = pyarrow.fs.FileSelector(
+        bucket_path, allow_not_found=True, recursive=False
+    )
+    return [
+        os.path.relpath(file_info.path.lstrip("/"), start=bucket_path.lstrip("/"))
+        for file_info in fs.get_file_info(selector)
+    ]
 
 
 def is_directory(uri: str) -> bool:
-    fs, fs_path = get_fs_and_path(uri)
-    if not fs:
-        raise ValueError(
-            f"URI `{uri}` is not a valid or supported cloud target. "
-            f"Hint: {fs_hint(uri)}"
-        )
+    """Checks if a remote URI is a directory or a file.
 
-    from ray.train._internal.fs_utils import _is_directory
+    Returns:
+        bool: True if the URI is a directory. False if it is a file.
 
-    return _is_directory(fs=fs, fs_path=fs_path)
+    Raises:
+        FileNotFoundError: if the URI doesn't exist.
+    """
+    _assert_pyarrow_installed()
+
+    fs, bucket_path = get_fs_and_path(uri)
+    file_info = fs.get_file_info(bucket_path)
+    return not file_info.is_file
 
 
 def _ensure_directory(uri: str, fs: Optional["pyarrow.fs.FileSystem"] = None):
-    fs, fs_path = get_fs_and_path(uri)
+    """Create directory at remote URI.
 
-    from ray.train._internal.fs_utils import _ensure_directory
+    Some external filesystems require directories to already exist, or at least
+    the `netloc` to be created (e.g. PyArrows ``mock://`` filesystem).
 
-    _ensure_directory(fs=fs, fs_path=fs_path)
+    Generally this should be done before and outside of Ray applications. This
+    utility is thus primarily used in testing, e.g. of ``mock://` URIs.
+
+    If a ``fs`` is passed, the ``uri`` is expected to be a path on this filesystem.
+    Otherwise, the fs and the uri are automatically detected.
+    """
+    if fs:
+        path = uri
+    else:
+        fs, path = get_fs_and_path(uri)
+    try:
+        fs.create_dir(path)
+    except Exception:
+        pass
